[#40] 기능: 안드로이드 갤러리 접근 후 파일 업로드 기능 추가
