+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.breadmap">
+    package="com.example.daedong">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
@@ -10,19 +10,20 @@
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <application
         android:allowBackup="true"
+        android:usesCleartextTraffic="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="false"
         android:hardwareAccelerated="true"
         android:requestLegacyExternalStorage="true"
-        android:name=".BreadmapApplication"
-        android:theme="@style/Theme.Breadmap">
+        android:name="com.example.daedong.DaedongApplication"
+        android:theme="@style/Theme.Daedong">
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.daedong.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Breadmap.NoActionBar">
+            android:theme="@style/Theme.Daedong.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
