--- android/app/src/main/java/com/example/breadmap/network/NetworkState.kt
@@ -1,10 +1,0 @@
-package com.example.breadmap.network
-
-sealed class Resource<T>(
-    val data: T? = null,
-    val message: String? = null
-) {
-    class Success<T>(data: T) : Resource<T>(data)
-    class Loading<T>(data: T? = null) : Resource<T>(data)
-    class Error<T>(message: String, data: T? = null) : Resource<T>(data, message)
-}
