--- android/app/src/main/java/com/example/breadmap/network/dto/AuthResponse.kt
@@ -1,5 +1,0 @@
-package com.example.breadmap.network.dto
-
-data class AuthResponse(
-    val appToken: String
-)
