--- android/app/src/main/java/com/example/breadmap/ui/theme/Color.kt
@@ -1,12 +1,0 @@
-package com.example.breadmap.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Primary = Color(0xFFFF6E40)
-val Secondary = Color(0xFFFFF6F4)
-
-val KakaoYellow = Color(0xFFFEE500)
-val KakaoLabel = Color(0xD9000000)
-
-val GoogleWhite = Color(0xFFFFFFFF)
-val GoogleLabel = Color(0xFF333333)
