--- android/app/src/main/java/com/example/breadmap/ui/theme/Shape.kt
@@ -1,13 +1,0 @@
-package com.example.breadmap.ui.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
-import androidx.compose.ui.unit.dp
-
-val ButtonRadiusShape = RoundedCornerShape(6.dp)
-
-val Shapes = Shapes(
-        small = RoundedCornerShape(4.dp),
-        medium = RoundedCornerShape(4.dp),
-        large = RoundedCornerShape(0.dp)
-)
