--- android/app/src/main/java/com/example/breadmap/utils/Constants.kt
@@ -1,7 +1,0 @@
-package com.example.breadmap.utils
-
-object Constants {
-    const val API_BASE_URI: String = "http://192.168.0.12:8080"
-    const val EXPIRED_TIME: Long = 2592000000L // millisecond
-    const val REFRESH_TIME: Long = 2160000000L // millisecond
-}
