--- android/app/src/main/java/com/example/breadmap/utils/GoogleSignInResultContract.kt
@@ -1,34 +1,0 @@
-package com.example.breadmap.utils
-
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import androidx.activity.result.contract.ActivityResultContract
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.tasks.Task
-import com.example.breadmap.R
-
-private fun getGoogleSignInClient(context: Context): GoogleSignInClient {
-    val signInOptions = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-        .requestIdToken(context.getString(R.string.google_key))
-        .requestEmail()
-        .build()
-
-    return GoogleSignIn.getClient(context, signInOptions)
-}
-
-class GoogleSignInResultContract : ActivityResultContract<Int, Task<GoogleSignInAccount>?>() {
-    override fun createIntent(context: Context, input: Int?): Intent {
-        return getGoogleSignInClient(context).signInIntent.putExtra("input", input)
-    }
-
-    override fun parseResult(resultCode: Int, intent: Intent?): Task<GoogleSignInAccount>? {
-        return when (resultCode) {
-            Activity.RESULT_OK -> GoogleSignIn.getSignedInAccountFromIntent(intent)
-            else -> null
-        }
-    }
-}
