+++ android/app/src/main/java/com/example/daedong/BreadmapApplication.kt
@@ -1,0 +1,12 @@
+package com.example.daedong
+
+import android.app.Application
+import com.kakao.sdk.common.KakaoSdk
+
+class DaedongApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        KakaoSdk.init(this, resources.getString(R.string.kakao_key))
+    }
+}
