+++ android/app/src/main/java/com/example/daedong/network/NetworkConstants.kt
@@ -1,0 +1,5 @@
+package com.example.daedong.network
+
+object NetworkConstants {
+    const val API_BASE_URI: String = "http://192.168.0.12:8080"
+}
