+++ android/app/src/main/java/com/example/daedong/network/NetworkState.kt
@@ -1,0 +1,10 @@
+package com.example.daedong.network
+
+sealed class Resource<T>(
+    val data: T? = null,
+    val message: String? = null
+) {
+    class Success<T>(data: T) : Resource<T>(data)
+    class Loading<T>(data: T? = null) : Resource<T>(data)
+    class Error<T>(message: String, data: T? = null) : Resource<T>(data, message)
+}
