+++ android/app/src/main/java/com/example/daedong/network/dto/AuthResponse.kt
@@ -1,0 +1,5 @@
+package com.example.daedong.network.dto
+
+data class AuthResponse(
+    val appToken: String
+)
