+++ android/app/src/main/java/com/example/daedong/ui/theme/Color.kt
@@ -1,0 +1,11 @@
+package com.example.daedong.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Primary500 = Color(0xFFFF6E40)
+
+val KakaoYellow = Color(0xFFFEE500)
+val KakaoLabel = Color(0xD9000000)
+
+val GoogleWhite = Color(0xFFFFFFFF)
+val GoogleLabel = Color(0xFF333333)
