+++ android/app/src/main/java/com/example/daedong/ui/theme/Theme.kt
@@ -1,0 +1,18 @@
+package com.example.daedong.ui.theme
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.darkColors
+import androidx.compose.material.lightColors
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DaedongTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    content: @Composable() () -> Unit
+) {
+    MaterialTheme(
+        colors = if (darkTheme) darkColors() else lightColors(),
+        content = content
+    )
+}
