+++ android/app/src/main/java/com/example/daedong/ui/utils/Constants.kt
@@ -1,0 +1,7 @@
+package com.example.daedong.ui.utils
+
+object Constants {
+    const val WEBVIEW_BASE_URI: String = "http://192.168.0.12:3000/addBread"
+    const val EXPIRED_TIME: Long = 2592000000L // millisecond
+    const val REFRESH_TIME: Long = 2160000000L // millisecond
+}
