+++ android/app/src/main/java/com/example/daedong/ui/utils/GoogleSignInResultContract.kt
@@ -1,0 +1,38 @@
+package com.example.daedong.ui.utils
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import androidx.activity.result.contract.ActivityResultContract
+import com.example.daedong.R
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.tasks.Task
+
+class GoogleSignInResultContract : ActivityResultContract<Int, Task<GoogleSignInAccount>?>() {
+    private fun getGoogleSignInClient(context: Context): GoogleSignInClient {
+        val signInOptions = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+            .requestIdToken(context.getString(R.string.google_key))
+            .requestEmail()
+            .build()
+
+        return GoogleSignIn.getClient(context, signInOptions)
+    }
+
+    override fun createIntent(context: Context, input: Int?): Intent {
+        val intent = getGoogleSignInClient(context).signInIntent.apply {
+            this.putExtra("input", input)
+        }
+
+        return intent
+    }
+
+    override fun parseResult(resultCode: Int, intent: Intent?): Task<GoogleSignInAccount>? {
+        return when (resultCode) {
+            Activity.RESULT_OK -> GoogleSignIn.getSignedInAccountFromIntent(intent)
+            else -> null
+        }
+    }
+}
