+++ android/app/src/main/java/com/example/daedong/ui/utils/UserSerializer.kt
@@ -1,0 +1,32 @@
+package com.example.daedong.ui.utils
+
+import android.content.Context
+import androidx.datastore.core.CorruptionException
+import androidx.datastore.core.DataStore
+import androidx.datastore.core.Serializer
+import androidx.datastore.dataStore
+import com.example.daedong.User
+import com.google.protobuf.InvalidProtocolBufferException
+import java.io.InputStream
+import java.io.OutputStream
+
+object UserSerializer : Serializer<User> {
+    override val defaultValue: User = User.getDefaultInstance()
+
+    override suspend fun readFrom(input: InputStream): User {
+        try {
+            return User.parseFrom(input)
+        } catch (exception: InvalidProtocolBufferException) {
+            throw CorruptionException("Cannot read proto", exception)
+        }
+    }
+
+    override suspend fun writeTo(t: User, output: OutputStream) {
+        t.writeTo(output)
+    }
+}
+
+val Context.userDataStore: DataStore<User> by dataStore(
+    fileName = "user.pb",
+    serializer = UserSerializer
+)
