+++ android/app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.Breadmap" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.Daedong" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/transparent</item>
         <item name="colorPrimaryVariant">@color/transparent</item>
