+++ android/app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.Breadmap" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.Daedong" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/transparent</item>
         <item name="colorPrimaryVariant">@color/transparent</item>
@@ -14,12 +14,12 @@
         <!-- Customize your theme here. -->
     </style>
 
-    <style name="Theme.Breadmap.NoActionBar">
+    <style name="Theme.Daedong.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
 
-    <style name="Theme.Breadmap.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+    <style name="Theme.Daedong.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
 
-    <style name="Theme.Breadmap.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+    <style name="Theme.Daedong.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 </resources>
